Move getAWSConfigFromProfile method to util
